Sprint 4 initial push.

+++ POS/src/main/java/com/ust/pos/customer/service/impl/CustomerServiceImpl.java
@@ -1,15 +1,16 @@
 package com.ust.pos.customer.service.impl;
 
 import com.ust.pos.address.service.impl.AddressServiceImpl;
+import com.ust.pos.commonservice.CommonService;
 import com.ust.pos.customer.service.CustomerService;
 import com.ust.pos.dto.AddressDto;
 import com.ust.pos.dto.CustomerDto;
 import com.ust.pos.dto.ShelfsDto;
+import com.ust.pos.dto.WsDto;
 import com.ust.pos.model.Customer;
 import com.ust.pos.model.CustomerRepository;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -18,16 +19,17 @@ import java.lang.reflect.Type;
 import java.util.List;
 
 @Service
-public class CustomerServiceImpl implements CustomerService {
+public class CustomerServiceImpl extends CommonService implements CustomerService {
 
-    @Autowired
-    private CustomerRepository customerRepository;
+    private final CustomerRepository customerRepository;
+    private final ModelMapper modelMapper;
+    private final AddressServiceImpl addressService;
 
-    @Autowired
-    private ModelMapper modelMapper;
-
-    @Autowired
-    private AddressServiceImpl addressService;
+    public CustomerServiceImpl(CustomerRepository customerRepository, ModelMapper modelMapper, AddressServiceImpl addressService) {
+        this.customerRepository = customerRepository;
+        this.modelMapper = modelMapper;
+        this.addressService = addressService;
+    }
 
     @Override
     public CustomerDto findByIdentifier(String identifier) {
@@ -49,12 +51,20 @@ public class CustomerServiceImpl implements CustomerService {
         String identifier = customerDto.getIdentifier();
         Customer existingCustomer = customerRepository.findByIdentifier(identifier);
         if (existingCustomer != null) {
+            if (existingCustomer.isDeleted()) {
+                customerDto.setMessage("Customer with identifier " + identifier + " was previously deleted. " +
+                        "Please contact backend team to restore."
+                );
+                customerDto.setSuccess(false);
+                return customerDto;
+            }
             customerDto.setMessage("Customer with identifier - " + identifier +
                     " already exists");
             customerDto.setSuccess(false);
             return customerDto;
         }
         Customer customer = modelMapper.map(customerDto, Customer.class);
+        setAuditFields(customer, true);
         customerRepository.save(customer);
 
         AddressDto billingAddress = customerDto.getBillingAddress();
@@ -82,6 +92,7 @@ public class CustomerServiceImpl implements CustomerService {
             return customerDto;
         }
         modelMapper.map(customerDto, existingCustomer);
+        setAuditFields(existingCustomer, false);
         customerRepository.save(existingCustomer);
         List<AddressDto> addresses = addressService.findAllByPhoneNo(customerDto.getIdentifier());
         AddressDto billingAddress = customerDto.getBillingAddress();
@@ -97,23 +108,34 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public boolean delete(String identifier) {
-        customerRepository.deleteByIdentifier(identifier);
+        Customer customer = customerRepository.findByIdentifier(identifier);
+        if (customer == null) return false;
+        softDelete(customer);
+        setAuditFields(customer, false);
+        customerRepository.save(customer);
         addressService.delete(identifier);
         return true;
     }
 
     @Override
-    public List<CustomerDto> findAll(Pageable pageable) {
+    public WsDto<CustomerDto> findAll(Pageable pageable) {
         Type listType = new TypeToken<List<CustomerDto>>() {
         }.getType();
-        Page<Customer> customerPage = customerRepository.findAll(pageable);
-        return modelMapper.map(customerPage.getContent(), listType);
+        Page<Customer> customerPage = customerRepository.findByDeletedFalse(pageable);
+        WsDto<CustomerDto> customerWsDto = new WsDto<>();
+        customerWsDto.setDtoList(modelMapper.map(customerPage.getContent(), listType));
+        customerWsDto.setTotalRecords(customerPage.getTotalElements());
+        customerWsDto.setTotalPages(customerPage.getTotalPages());
+        customerWsDto.setSizePerPage(pageable.getPageSize());
+        customerWsDto.setPage(pageable.getPageNumber());
+        return customerWsDto;
     }
 
     @Override
     public CustomerDto toggleStatus(String identifier) {
         Customer customer = customerRepository.findByIdentifier(identifier);
         customer.setStatus(!customer.isStatus());
+        setAuditFields(customer, false);
         customerRepository.save(customer);
         return modelMapper.map(customer, CustomerDto.class);
     }
@@ -122,6 +144,11 @@ public class CustomerServiceImpl implements CustomerService {
     public List<CustomerDto> findIfTrue() {
         Type listType = new TypeToken<List<ShelfsDto>>() {
         }.getType();
-        return modelMapper.map(customerRepository.findByStatusIsTrue(), listType);
+        return modelMapper.map(customerRepository.findByStatusIsTrueAndDeletedFalse(), listType);
+    }
+
+    @Override
+    public CustomerDto findByIdentifierAndDeletedFalse(String identifier) {
+        return modelMapper.map(customerRepository.findByIdentifierAndDeletedFalse(identifier), CustomerDto.class);
     }
 }
